feat(toggl): Add multiple task mode
Multiple task mode chooses how to handle the case that multiple tasks are detected for a single time entry. Whether to evenly divide the time up or just assign the time to the first task.

Also allowed task detection regex to contain groups to define what part of the match is the task name.